requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
packaging==23.0
Pillow==9.4.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
Wand==0.6.11
